Updated file check and asset path in Firefox release workflow
The changes involve updating the filename that the script checks for in the web-ext-artifacts directory. Previously, it was looking for any zip file with a name starting with 'tonscanplus-', but now it specifically looks for 'tonscan_-1.1.zip'. The same change has been made to the asset_path used when uploading to GitHub.